pyproject.toml
[project]
name = "projecteweb"
version = "0.1.0"
description = ""
authors = [
    {name = "caarlos-04",email = "cmazarico@gmail.com"},
    {name = "oriolvines", email = "oriolvibo04@gmail.com"},
    {name = "David0-Eng", email = "davidmf12004@gmail.com"},
    {name = "blaigene", email = "genemorablai@gmail.com"},
    {name = "polma7", email = "polmasip7@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "behave-django",
    "spotipy (>=2.25.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
